pyproject.toml
[tool.poetry]
name = "core"
version = "2"
description = ""
authors = ["Tony <tony.lbvre@gmail.com>"]
packages = [
    { include = "core" },
]

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^2.11.3"
Babel = "^2.9.0"
python-box = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
